requirements.txt
# Core application
flask==2.3.3
python-dotenv==1.0.0

# Database
flask-sqlalchemy==3.0.5

# Instagram API interaction
requests==2.31.0
requests-oauthlib==1.3.1

# Image manipulation
pillow==10.0.0

# Content generation
transformers==4.30.2
torch==2.0.1

# Background tasks & scheduling
apscheduler==3.10.1

# Development tools
pytest==7.4.0
flake8==6.1.0